Unbreak cartographer_turtlebot and cartographer_toyota_hsr. (#225)
FindThreads.cmake (e.g. of trusty) is used which adds the
"-lpthead" dependency that is added to cartographer's
dependencies. Catkin then fails to resolve this in the
cartographer_rosConfig.cmake file it generates to be used
by all dependencies of cartographer_ros.

This means for now, users of cartographer_ros also have to
directly depend on cartographer.